forth commit message

+++ src/Controller/EmployeesController.php
@@ -1,7 +1,6 @@
 <?php
-namespace App\Controller;
 
-use App\Controller\AppController;
+namespace App\Controller;
 
 /**
  * Employees Controller
@@ -17,50 +16,54 @@ class EmployeesController extends AppController
      *
      * @return \Cake\Http\Response|null
      */
+
+    public function initialize(): void
+    {
+        parent::initialize();
+        $this->loadModel('Departments');// loads table
+        $this->loadModel('Designations');
+    }
+
+
     public function index()
     {
-      // Get search keyword from URL
-      $search = $this->request->getQuery('search');
-      $department = $this->request->getQuery('department');
+        // Get search keyword from URL
+      $search = $this->request->getQuery('search');//gets form the input from the form
+      $department = $this->request->getQuery('department');//and storesit in the variable
       $status = $this->request->getQuery('status');
 
-      // Start query
-      $query = $this->Employees->find()
-        ->contain(['Departments', 'Designations']);
+        // Start query
+        $query = $this->Employees->searchEmployees($search, $department, $status);
 
-     // Apply search if user entered something
-     if (!empty($search)) {
+        // // Apply search if user entered something
+        // if (!empty($search)) {
+        //     $query->where([
+        //     'OR' => [
+        //         'Employees.employee_code LIKE' => '%' . $search . '%',
+        //         'Employees.name LIKE' => '%' . $search . '%',
+        //         'Employees.email LIKE' => '%' . $search . '%',
+        //         'Employees.mobile LIKE' => '%' . $search . '%',
+        //     ]
+        // ]);
+        // }
+        // //4 deaprtment filter
+        // if (!empty($department)) {
+        //     $query->where(['Employees.department_id' => $department]);
+        // }
 
-        $query->where([
-            'OR' => [
-                'Employees.employee_code LIKE' => '%' . $search . '%',
-                'Employees.name LIKE' => '%' . $search . '%',
-                'Employees.email LIKE' => '%' . $search . '%',
-                'Employees.mobile LIKE' => '%' . $search . '%',
-            ]
-        ]);
-     }
-     //4 deaprtment filter
-    if (!empty($department)) 
-        {
-         $query->where(['Employees.department_id' => $department]);
-        }
+        // if (!empty($status)) {
+        //     $query->where(['Employees.status' => $status]);
+        // }
 
-    if (!empty($status))
-    {
-      $query->where(['Employees.status' => $status]);
-    }
-    // ⭐ 5. LOAD DEPARTMENTS HERE ⭐
-    $departments = $this->Employees
-        ->Departments
-        ->find('list')
-        ->order(['name' => 'ASC']);
+        // 5. LOAD DEPARTMENTS HERE
+        $departments = $this->Departments->getDepartmentList();
+        $designations = $this->Designations->getDesignationList();
 
-    // 6. Pagination  
-    $this->paginate = ['limit' => 10, 'order'=>['Employees.name'=>'ASC']];
-    $this->request = $this->request->withAttribute('paging',$this->request->getAttribute('paging'));
-    $employees = $this->paginate($query);
-    $this->set(compact('employees','search','department','departments','status'));
+        // 6. Pagination
+        $this->paginate = ['limit' => 10, 'order'=>['Employees.name'=>'ASC']];
+        $this->request = $this->request->withAttribute('paging', $this->request->getAttribute('paging'));
+        $employees = $this->paginate($query);
+        $this->set(compact('employees', 'search', 'department', 'departments', 'status'));
     }
 
     /**
@@ -72,10 +75,7 @@ class EmployeesController extends AppController
      */
     public function view($id = null)// called when /employees/view/5
     {
-        $employee = $this->Employees->get($id, [
-            'contain' => ['Departments', 'Designations', 'Attendances', 'Payslips'],
-            ]);// access the other tables too using orm
-
+        $employee =$this->Employees->getEmployeedetails($id);
         $this->set('employee', $employee);
     }
 
@@ -90,19 +90,15 @@ class EmployeesController extends AppController
         if ($this->request->is('post')) {
             $employee = $this->Employees->patchEntity($employee, $this->request->getData());// it trigers the employee entity wiht the post data and checks valiadtion roles too
             if ($this->Employees->save($employee)) {
-           $this->Flash->success(__('Employee added successfully.'));
-             return $this->redirect(['action' => 'index']);
-
-             } else {
-
-             debug($employee->getErrors());
-             die();
-
+                $this->Flash->success(__('Employee added successfully.'));
+                return $this->redirect(['action' => 'index']);
+            } else {
+                debug($employee->getErrors());
+                die();
             }
-      
-            }
-        $departments = $this->Employees->Departments->find('list', ['limit' => 200]);
-        $designations = $this->Employees->Designations->find('list', ['limit' => 200]);
+        }
+        $departments = $this->Departments->getDepartmentList();
+        $designations = $this->Designations->getDesignationList();
         $this->set(compact('employee', 'departments', 'designations'));
     }
 
@@ -127,8 +123,8 @@ class EmployeesController extends AppController
             }
             $this->Flash->error(__('The employee could not be saved. Please, try again.'));
         }
-        $departments = $this->Employees->Departments->find('list', ['limit' => 200]);
-        $designations = $this->Employees->Designations->find('list', ['limit' => 200]);
+        $departments = $this->Departments->getDepartmentList();
+        $designations = $this->Designations->getDesignationList();
         $this->set(compact('employee', 'departments', 'designations'));
     }
 
@@ -144,14 +140,13 @@ class EmployeesController extends AppController
         $this->request->allowMethod(['post', 'delete']);// deletion can be from post or delete req
         $employee = $this->Employees->get($id, ['contain' => ['Attendances', 'Payslips']]);
 
-        if (!empty($employee->attendances) ||!empty($employee->payslips)) {
-              $this->Flash->error(__('Employee cannot be deleted because attendance or payroll records exist.'));
-                return $this->redirect(['action' => 'index']);
-            }
+        if (!$this->Employees->canDeleteEmployee($id)) {
+            $this->Flash->error(__('Employee cannot be deleted because attendance or payroll records exist.'));
+            return $this->redirect(['action' => 'index']);
+        }
         if ($this->Employees->delete($employee)) {
             $this->Flash->success(__('The employee has been deleted.'));
-        } 
-        else {
+        } else {
             $this->Flash->error(__('The employee could not be deleted. Please, try again.'));
         }
